don't try and call Object#id when calling Mash#id
Ruby 1.8 has an Object#id method that is called when the
Mash doesn't have an id key. This causes unexpected failures since
there is no id key-value, but Mash is still returning a value. If the
object_id is needed just use Object#object_id or Object#__id__

(Same with type -> class)